✨ LumaFlow: Refactor from LumeFlow + Fix UV plane bug
- Renamed project from LumeFlow to LumaFlow for consistency
- Fixed critical UV plane bug causing green artifacts in depth data
- Added UV plane initialization (128,128) for proper grayscale depth
- Normalized timestamps for correct video duration
- Fixed camera preview orientation
- Added Files app support for easy video access
- Added depth analysis tools (analyze_depth.py)
- Documented UV plane fix and depth quality improvements
- Added training data progress tracker

Changes:
- LumaFlow/ - Complete refactored iOS app
- FileWriter.swift - UV plane fix for clean depth encoding
- DEPTH_QUALITY_FIX.md - Depth quality guide
- GREEN_ARTIFACTS_FIX.md - UV plane bug documentation
- TRAINING_DATA_PROGRESS.md - Training data tracker
- analyze_depth.py - Python tool for depth visualization

Status: Ready for training data collection (2/15 videos captured)

+++ LumaFlow/LumaFlow/Services/OnDeviceEncoder.swift
@@ -195,7 +195,7 @@ class OnDeviceEncoder {
         // Header
         let header = LumaFlowHeader(
             version: 1,
-            frameCount: encodedFrames.count,
+            frameCount: UInt32(encodedFrames.count),
             width: 1920,
             height: 1080,
             frameRate: 30
